# Reduce number of regexes for performance