--- a/lib/normalizer.js
+++ b/lib/normalizer.js
@@ -43,7 +43,7 @@ var readSubstitutes = function(file, lineHandle, closeHandle) {
 
     // Lets allow comments with '#'
     var pos = nline.indexOf('#');
-    
+
     if (pos === -1) {
       var parts = nline.s.split(" ");
 
@@ -53,7 +53,7 @@ var readSubstitutes = function(file, lineHandle, closeHandle) {
 
         lineHandle(parts[0], parts[1]);
       }
-      
+
     } else if (pos > 0) {
       nline = nline.left(pos);
       var parts = nline.s.split(" ");
@@ -75,10 +75,10 @@ exports.loadData = function(cb){
       if (reSet[item.key] === undefined) {
         reSet[item.key] = {};
       }
-      
+
       if (reSet[item.key][key] === undefined) {
         reSet[item.key][key] = [];
-      
+
         // Add RegEx
         var startM, endM, lookup = key;
         if (key[0] == '<') {
@@ -96,25 +96,20 @@ exports.loadData = function(cb){
         if (startM && endM) {
           reSet[item.key][key].push({re: new RegExp("^" + qm + "$", "gi"), r: replacer });
         } else if (startM) {
-          reSet[item.key][key].push({re: new RegExp("^" + qm + "$", "gi"), r: replacer });
-          reSet[item.key][key].push({re: new RegExp("^" + qm + "(\\W+)", "gi"), r: replacer + "$1"});
+          reSet[item.key][key].push({re: new RegExp("^" + qm + "(\\W+|$)", "gi"), r: replacer + "$1"});
         } else if (endM) {
-
-          reSet[item.key][key].push({re: new RegExp("(\\W+)" + qm + "$", "gi"), r: "$1" + replacer });
+          reSet[item.key][key].push({re: new RegExp("(\\W+|^)" + qm + "$", "gi"), r: "$1" + replacer });
           if (item.key == "_sys") {
             reSet[item.key][key].push({re: new RegExp(qm + "$", "gi"), r: replacer });
           } else {
             // reSet[item.key][key].push({re: new RegExp("(\\W+)" + qm + "(\\W+)", "gi"), r: "$1" + replacer + "$2" });
           }
         } else {
-          reSet[item.key][key].push({re: new RegExp("^" + qm + "$", "gi"), r: replacer });
-          reSet[item.key][key].push({re: new RegExp("^" + qm + "(\\W+)", "gi"), r: replacer + "$1" });
-          reSet[item.key][key].push({re: new RegExp("(\\W+)" + qm + "(\\W+)", "gi"), r: "$1" + replacer + "$2" });
-          reSet[item.key][key].push({re: new RegExp("(\\W+)" + qm + "$", "gi"), r: "$1" + replacer });
+          reSet[item.key][key].push({re: new RegExp("(\\W+|^)" + qm + "(\\W+|$)", "gi"), r: "$1" + replacer + "$2" });
         }
       }
     };
-    
+
     readSubstitutes(item.file, lineHandle, function() {
       fc++;
       if (tasks.length == fc) {