--- a/test/basic.js
+++ b/test/basic.js
@@ -78,4 +78,84 @@ describe('Normalizer', function(){
       norm.clean("œ").should.eql('');
     });
   });
+
+  describe('Matching', function() {
+    // <it_is>
+    describe('<xxx>', function() {
+      it('should match start and end', function() {
+        norm.clean('it is').should.eql("~yes");
+      });
+
+      it('should not match start', function() {
+        norm.clean('it is abc').should.eql("it is abc");
+      });
+
+      it('should not match end', function() {
+        norm.clean('abc it is').should.eql('abc it is');
+      });
+
+      it('should not match middle', function() {
+        norm.clean('abc it is abc').should.eql('abc it is abc');
+      });
+    });
+
+    // <ew
+    describe('<xxx', function() {
+      it('should match start and end', function() {
+        norm.clean('ew').should.eql("~emodisgust");
+      });
+
+      it('should match start', function() {
+        norm.clean('ew abc').should.eql("~emodisgust abc");
+      });
+
+      it('should not match end', function() {
+        norm.clean('abc ew').should.eql("abc ew");
+      });
+
+      it('should not match middle', function() {
+        norm.clean('abc ew abc').should.eql("abc ew abc");
+      });
+    });
+
+    // have_to_go>
+    describe('xxx>', function() {
+      it('should match start and end', function() {
+        norm.clean('have to go').should.eql("~emogoodbye");
+      });
+
+      it('should not match start', function() {
+        norm.clean('have to go abc').should.eql("have to go abc");
+      });
+
+      it('should match end', function() {
+        norm.clean('abc have to go').should.eql("abc ~emogoodbye")
+      });
+
+      it('should not match middle', function() {
+        norm.clean('abc have to go abc').should.eql("abc have to go abc")
+      });
+    });
+
+    // okay
+    describe('xxx', function() {
+      it('should match start and end', function() {
+        norm.clean('okay').should.eql("~yes");
+      });
+
+      it('should match start', function() {
+        norm.clean('okay abc').should.eql("~yes abc");
+      });
+
+      it('should match end', function() {
+        norm.clean('abc okay').should.eql("abc ~yes");
+      });
+
+      it('should match middle', function() {
+        norm.clean('abc okay abc').should.eql("abc ~yes abc");
+      });
+    })
+
+
+  });
 });