And add tests for matching.

Before the new tests were added, here’s the timing for three test separate test runs before/after the update:

Before:
2388ms
2429ms
2515ms

After:
1133ms
1124ms
1673ms